Extract buildBaseTypes + helpers into scripts/buildBaseTypes.js
Addresses review comment from
https://github.com/pmossman/forceteki/pull/1#discussion_r3215219340
(option chosen: extract to its own module).

scripts/fetchdata.js had grown by ~125 lines of base-type-building
logic (buildBaseTypes + labelForGroup + formatRaritySuffix +
capitalizeWord + slug helpers). Moves them out into a dedicated
scripts/buildBaseTypes.js module that exports buildBaseTypes;
fetchdata.js gains a single require + the existing one-line call.

Net behavior unchanged. fetchdata.js is back to focusing on card-data
fetch/parse (534 lines, was 659); buildBaseTypes.js owns the grouping
concern in isolation (125 lines).
